+++ Tests/FoundationTests/NSPredicateExtensionsTests.swift
@@ -39,4 +39,41 @@ final class NSPredicateExtensionsTests: XCTestCase {
             XCTAssertEqual(subpredicates, [predicate1, predicate2])
         }
     }
+    
+    func testOperatorNot() {
+        let predicate = NSPredicate(format: "a < 7")
+        let notPredicate = !predicate
+        XCTAssert(notPredicate.compoundPredicateType == .not)
+        if let subpredicates = notPredicate.subpredicates as? [NSPredicate] {
+            XCTAssertEqual(subpredicates, [predicate])
+        }
+    }
+    
+    func testOperatorAndPredicate() {
+        let predicate1 = NSPredicate(format: "a < 7")
+        let predicate2 = NSPredicate(format: "a > 3")
+        let andPredicate = predicate1 + predicate2
+        XCTAssert(andPredicate.compoundPredicateType == .and)
+        if let subpredicates = andPredicate.subpredicates as? [NSPredicate] {
+            XCTAssertEqual(subpredicates, [predicate1, predicate2])
+        }
+    }
+    
+    func testOperatorOrPredicate() {
+        let predicate1 = NSPredicate(format: "a < 7")
+        let predicate2 = NSPredicate(format: "a > 3")
+        let orPredicate = predicate1 | predicate2
+        XCTAssert(orPredicate.compoundPredicateType == .or)
+        if let subpredicates = orPredicate.subpredicates as? [NSPredicate] {
+            XCTAssertEqual(subpredicates, [predicate1, predicate2])
+        }
+    }
+    
+    func testOperatorSubPredicate() {
+        let predicate1 = NSPredicate(format: "SELF BETWEEN{1,5}")
+        let predicate2 = NSPredicate(format: "SELF BETWEEN{3,6}")
+        let subPredicate = predicate1 - predicate2
+        XCTAssert(subPredicate.evaluate(with: 2))
+        XCTAssertFalse(subPredicate.evaluate(with: 4))
+    }
 }